move: style 파일 전체 분리
