+++ src/pages/ChatPage/ChatRoom.jsx
@@ -1,14 +1,14 @@
 import React from 'react';
 import FeedHeader from '../../components/common/Header/FeedHeader';
 import ChatRoomContents from './ChatRoomContents';
-import Comment from './Comment';
+import ChatComment from './ChatComment';
 
 const ChatList = () => {
   return (
     <>
       <FeedHeader isChatHeader={true} />
       <ChatRoomContents />
-      <Comment atChatroom={true} />
+      <ChatComment atChatroom={true} />
     </>
   );
 };
