+++ src/pages/JoinPage/ProfileSetting.jsx
@@ -1,37 +1,9 @@
 import React, { useEffect, useState } from 'react';
-import styled from 'styled-components';
-import Wrapper from '../../components/common/Wrapper/Wrapper';
-import Button from '../../components/common/Button/Button';
+import * as S from './ProfileSetting.style';
 import ProfileForm from '../../components/Profile/ProfileForm';
 import { useLocation, useNavigate } from 'react-router-dom';
 import { join } from '../../api/auth';
 import useImageUpload from '../../hooks/useImageUpload';
-
-const Main = styled.main``;
-
-const LayoutWrapper = styled(Wrapper)`
-  padding: 0 34px;
-`;
-
-const Form = styled.form``;
-
-const Title = styled.h1`
-  margin: 30px 0 12px;
-  font-size: 24px;
-  font-weight: 500;
-  text-align: center;
-`;
-
-const Desc = styled.p`
-  font-size: 14px;
-  text-align: center;
-  color: ${({ theme }) => theme.colors.txtColor};
-  margin-bottom: 30px;
-`;
-
-const StartBtn = styled(Button)`
-  margin-top: 30px;
-`;
 
 const ProfileSetting = () => {
   const [formData, setFormData] = useState({
@@ -56,29 +28,6 @@ const ProfileSetting = () => {
     if (!state?.email || !state?.password) navigate('/join');
   }, []);
 
-  // 프로필 수정 페이지에서 버튼이 form 밖에 있어서 onSubmit으로 하지 않고 onClick으로 구현
-  // const handleClick = async () => {
-  //   try {
-  //     const image = img === '' ? 'https://api.mandarin.weniv.co.kr/Ellipse.png' : img;
-
-  //     const response = await fetch('https://api.mandarin.weniv.co.kr/user', {
-  //       method: 'POST',
-  //       headers: {
-  //         'Content-Type': 'application/json',
-  //       },
-  //       body: JSON.stringify({ user: { ...state, ...formData, image } }),
-  //     });
-
-  //     const data = await response.json();
-  //     if (data.message === '회원가입 성공') {
-  //       navigate('/');
-  //     } else {
-  //       navigate('/join');
-  //     }
-  //   } catch (err) {
-  //     console.log(err.message);
-  //   }
-  // };
   const handleClick = async () => {
     try {
       const image = img === '' ? 'https://api.mandarin.weniv.co.kr/Ellipse.png' : img;
@@ -95,11 +44,11 @@ const ProfileSetting = () => {
   };
 
   return (
-    <Main>
-      <LayoutWrapper>
-        <Form onSubmit={(e) => e.preventDefault()}>
-          <Title>프로필 설정</Title>
-          <Desc>나중에 얼마든지 변경할 수 있습니다.</Desc>
+    <S.Main>
+      <S.LayoutWrapper>
+        <S.Form onSubmit={(e) => e.preventDefault()}>
+          <S.Title>프로필 설정</S.Title>
+          <S.Desc>나중에 얼마든지 변경할 수 있습니다.</S.Desc>
           <ProfileForm
             formData={formData}
             setFormData={setFormData}
@@ -108,12 +57,12 @@ const ProfileSetting = () => {
             error={error}
             setError={setError}
           />
-          <StartBtn mode={isActivated ? 'default' : 'disabled'} size='lg' onClick={handleClick}>
+          <S.StartBtn mode={isActivated ? 'default' : 'disabled'} size='lg' onClick={handleClick}>
             감귤마켓 시작하기
-          </StartBtn>
-        </Form>
-      </LayoutWrapper>
-    </Main>
+          </S.StartBtn>
+        </S.Form>
+      </S.LayoutWrapper>
+    </S.Main>
   );
 };
